requirements.txt
bip-utils==2.3.0
django==5
anchorpy==0.15.0
djangorestframework==3.15.2
djangorestframework-camel-case==1.3.0
dj-database-url==0.5.0
django-celery-results==2.3.1
django-bmemcached==0.3.0
django-cors-headers==3.12.0
django-encrypted-model-fields==0.6.5
django-filter==21.1
django-polymorphic==3.1.0
django-rest-polymorphic==0.1.10
drf-yasg==1.20.0
ed25519==1.5
psycopg2-binary==2.9.3
gunicorn==20.1.0
locust==2.26.0
redis==4.6
python-dotenv==0.20.0
python-memcached==1.59
sentry-sdk==1.5.12
uWSGI==2.0.22
web3==6.9.0
whitenoise==6.1.0
solders==0.14.3
coverage~=7.3.2
pytz~=2023.3.post1
requests~=2.31.0
celery~=5.3.4
celery-redbeat==2.2.0
django-safedelete~=1.3.3
tweepy~=4.14.0
pillow==10.4.0
channels[daphne]==4.1
channels_redis==4.2
